+++ src/assets/thumbnail/logo.png
